pyproject.toml
[tool.poetry]
name = "toja"
version = "1.0.0b5"
description = "The Old Joke Archive"
authors = ["Mark Hall <mark.hall@work.room3b.eu>"]
license = "MIT"
include = ["toja/server/frontend/public/build/*.*"]
exclude = ["toja/server/frontend/src", "toja/server/frontend/*.*"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.0"
cerberus = "^1.3.2"
email-validator = "^1.1.3"
pillow = "^8.1.1"
tesserocr = "^2.5.1"
PyYAML = "^5.4.1"
tornado = "^6.1"
aiocouch = "^2.1.0"
bcrypt = "^3.2.0"
Pillow = "^8.4.0"
asyncio-mqtt = "^0.12.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
flake8-print = "^3.1.4"
pre-commit = "^2.7.1"
tox = "^3.20.1"
flake8-docstrings = "^1.6.0"
flake8-annotations = "^2.6.2"
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
aiohttp = "^3.8.1"
Jinja2 = "^3.0.0"

[tool.poetry.extras]
postgresql = ["psycopg2"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
